requirements.txt
youtube-transcript-api
langchain-community
langchain-openai
langchain-google
faiss-cpu
tiktoken
python-dotenv